requirements.txt
streamlit==1.34.0
pandas==2.2.2
plotly==5.21.0
torch==2.3.0
sentence-transformers==2.6.1
